Implement Archive type for versioned parsing
Instead of needing to pass the file_version everywhere, extract it into
an Archive type that combines a reader and a version, and pass that
instead of the reader.

This means I can use the same deserializer for two different versions,
which'll be needed to parse the exports, considering how much versioning
exists in the FObjectExport struct serialization.